Work done
This branch implements the hashmap versions of the follower/following screens. Introduces a new variable called global.currentUser that can be used throughout the screens. I also changed the server to the one Pantea created so that the fetches work. There is a User object that is created for each of the users in the following_users array and followers array in the database, as well as the current user who is using the app. To construct the user, put values in an object. and send it to the User constructor. The User class can also get and set values. The follower/following screens only work when you sign in and not sign up because the post request is not working at the moment to add a user to the database. Once the post request works, the process for the follower/following screens to work should be the same as just signing in. This version does not save the user after exiting out of the emulator or reloading the app. This version has the very start of the search screen, meaning imports and class declaration.