Search Screen and Minor Adjustments
Made different screens that correspond to different tabs in the search screen. Adjusted minor things. The text is saved when typing into the search bar, have not implemented showing results yet. Changed the searchbar to a touchable highlight on the hometrip screen because it was hard to make the search bar go to a different screen.